+++ frontend/src/js/app.js
@@ -3,6 +3,8 @@ import { initDocsEnhancements } from "./modules/docs.js";
 import { initModals } from "./modules/modals.js";
 import { initMessages } from "./modules/messages.js";
 import { initNavigationShortcuts } from "./modules/shortcuts.js";
+import { initRepositoryLikes } from "./modules/repository-likes.js";
+import { showSourceEasterEgg } from "./modules/source-easter-egg.js";
 import { initTheme } from "./modules/theme.js";
 import { initUserSettingsCache } from "./modules/user-settings.js";
 
@@ -11,7 +13,9 @@ document.addEventListener("DOMContentLoaded", () => {
   initModals();
   initNavigationShortcuts();
   initMessages();
+  initRepositoryLikes();
   initCopyButtons();
   initDocsEnhancements();
   initUserSettingsCache();
+  showSourceEasterEgg();
 });